safari-600.7-branch/LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.3_Line_Terminators/S7.3_A3.1_T3-expected.txt

CONSOLE MESSAGE: line 77: SyntaxError: Unexpected identifier 'comment'
S7.3_A3.1_T3

PASS Expected parsing failure

TEST COMPLETE
